pyproject.toml
[build-system]
requires = ["setuptools>=75.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "vireo"
version = "0.2.7"
description = "AI-powered wildlife photo organizer"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.3",
    "huggingface_hub>=1.4.1",
    "imagehash>=4.3.2",
    "lightning>=2.6.1",
    "numpy>=2.4.2",
    "omegaconf>=2.3.0",
    "Pillow>=12.1.1",
    "pybioclip>=2.1.1",
    "PytorchWildlife>=1.2.4.2",
    "rawpy>=0.26.1",
    "scipy>=1.17.1",
    "transformers>=5.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=9.0",
    "pytest-cov>=6.0",
    "ruff>=0.11",
]

[tool.setuptools.packages.find]
include = ["vireo*"]

[tool.pytest.ini_options]
testpaths = ["tests", "vireo/tests"]

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "F",    # pyflakes
    "UP",   # pyupgrade
    "B",    # flake8-bugbear
    "SIM",  # flake8-simplify
    "I",    # isort
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "E741",   # ambiguous variable name
    "B010",   # do not call setattr with a constant
    "SIM108", # ternary operator
    "UP031",  # printf-style formatting
]

[tool.ruff.lint.per-file-ignores]
"vireo/tests/*" = ["E402", "B011", "F841", "SIM115", "SIM117"]
"tests/*" = ["E402", "B017", "F841"]
# -- production code: pre-existing violations, fix in dedicated cleanup PR --
"vireo/app.py" = ["B007", "F841", "SIM105", "SIM118"]
"vireo/classifier.py" = ["B905"]
"vireo/classify_job.py" = ["F821", "SIM105"]
"vireo/culling.py" = ["B007", "F841", "SIM105"]
"vireo/detector.py" = ["B904"]
"vireo/dino_embed.py" = ["B904"]
"vireo/importer.py" = ["B007"]
"vireo/jobs.py" = ["SIM105"]
"vireo/masking.py" = ["B904"]
"vireo/models.py" = ["B904", "F401"]
"vireo/quality.py" = ["B905"]
"vireo/selection.py" = ["F841"]
"vireo/sharpness.py" = ["SIM105", "SIM118"]
"vireo/taxonomy.py" = ["B905"]
# -- test files: violations not covered by vireo/tests/* glob --
"vireo/tests/test_culling.py" = ["E702", "E731"]
"vireo/tests/test_detector.py" = ["F401"]
"vireo/tests/test_dinov2.py" = ["B007"]
